refactor: move UpdateContribution to separate component
