+++ src/modules/contribution/ContributionPage.vue
@@ -12,56 +12,7 @@
 
       <ContributionBox :contribution="currentContribution" class="mb-9" />
 
-      <form class="mb-7 md:mb-12" @submit.prevent="submitContribution">
-        <SectionTitle class="mb-2"
-          >{{ t('contribution.billing') }}
-        </SectionTitle>
-
-        <p v-if="hasManualType" class="mb-4">
-          {{ t('contribution.manualPayment') }}
-        </p>
-
-        <Contribution
-          v-model:amount="newContribution.amount"
-          v-model:payFee="newContribution.payFee"
-          v-model:period="newContribution.period"
-          v-model:paymentMethod="newContribution.paymentMethod"
-          :content="contributionContent"
-          :show-period="showChangePeriod"
-          :show-payment-method="!isAutoActiveMember"
-        />
-
-        <ProrateContribution
-          v-if="showProrateOptions"
-          v-model="newContribution.prorate"
-          :new-amount="newContribution.amount"
-          :old-amount="currentContribution.amount!"
-          :renewal-date="currentContribution.renewalDate!"
-        />
-
-        <MessageBox v-if="hasUpdatedContribution" class="mb-4" type="success">
-          {{ t('contribution.updatedContribution') }}
-        </MessageBox>
-
-        <MessageBox v-if="cantUpdateContribution" class="mb-4" type="error">
-          {{ t('contribution.contributionUpdateError') }}
-        </MessageBox>
-
-        <AppButton
-          :disabled="!canSubmitContribution || validation.$invalid"
-          type="submit"
-          variant="link"
-          class="mb-4 w-full"
-          :loading="submitContributionLoading"
-        >
-          {{ contributionButtonText }}
-        </AppButton>
-
-        <InfoMessage
-          v-if="!isAutoActiveMember"
-          :message="t('contribution.changeBankInfo')"
-        />
-      </form>
+      <UpdateContribution />
 
       <PaymentSource
         v-if="currentContribution.paymentSource"
@@ -82,7 +33,6 @@
 import { computed, onBeforeMount } from 'vue';
 import { useI18n } from 'vue-i18n';
 import { useRoute } from 'vue-router';
-import useVuelidate from '@vuelidate/core';
 import { useContribution } from './use-contribution';
 
 import ContributionBox from './components/ContributionBox.vue';
@@ -90,15 +40,10 @@ import CancelContribution from './components/CancelContribution.vue';
 import PaymentSource from './components/PaymentSource.vue';
 
 import PageTitle from '../../components/PageTitle.vue';
-import InfoMessage from '../../components/InfoMessage.vue';
-import SectionTitle from '../../components/SectionTitle.vue';
-import Contribution from '../../components/contribution/Contribution.vue';
-import AppButton from '../../components/forms/AppButton.vue';
-import ProrateContribution from './components/ProrateContribution.vue';
 import AppAlert from '../../components/AppAlert.vue';
-import MessageBox from '../../components/MessageBox.vue';
 import PaymentsHistory from './components/PaymentsHistory.vue';
 import { currentUser } from '../../store';
+import UpdateContribution from './components/UpdateContribution.vue';
 
 const { t } = useI18n();
 
@@ -106,29 +51,12 @@ const route = useRoute();
 const updatedPaymentSource = route.query.updatedPaymentSource !== undefined;
 const startedContribution = route.query.startedContribution !== undefined;
 
-const validation = useVuelidate();
-
 const email = computed(() =>
   currentUser.value ? currentUser.value.email : ''
 );
 
-const {
-  isIniting,
-  initContributionPage,
-  currentContribution,
-  newContribution,
-  contributionContent,
-  canSubmitContribution,
-  submitContribution,
-  submitContributionLoading,
-  cantUpdateContribution,
-  hasUpdatedContribution,
-  showChangePeriod,
-  showProrateOptions,
-  contributionButtonText,
-  hasManualType,
-  isAutoActiveMember,
-} = useContribution();
+const { isIniting, initContributionPage, currentContribution } =
+  useContribution();
 
 onBeforeMount(() => {
   initContributionPage();
